% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucscHubgeneratorExample.R
\name{downloadUcschubgenerator}
\alias{downloadUcschubgenerator}
\title{Download ucschubgenerator example datasets}
\usage{
downloadUcschubgenerator(
  output_dir = "./",
 
    file_dir = "http://gattaca.imppc.org/genome_browser/lplab/exampleHub/example_data.tar.gz"
)
}
\arguments{
\item{output_dir}{Output directory for the datasets.}

\item{file_dir}{Path to the compressed database file.}
}
\value{
It creates the \code{output_dir} with the example ucschubgenerator files used
 by the vignette.
}
\description{
Downloads the required ucschubfenerator example datasets.
}
\examples{
path <- downloadUcschubgenerator()
}
