% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsyncHub.R
\name{rsyncHub}
\alias{rsyncHub}
\title{rsyncHub}
\usage{
rsyncHub(
  path_local_hub,
  hub_name,
  gattaca_html = "http://gattaca.imppc.org/genome_browser/lplab",
  gattaca_dir = "/data/apache/htdocs/genome_browser/lplab",
  gattaca_folder_hub,
  gattaca_user
)
}
\arguments{
\item{path_local_hub}{Local folder of the hub where the files generated will be saved. A folder with the \code{hub_name}
will be generated.}

\item{hub_name}{Single-word name of the directory containing the track hub files. Not displayed to hub users.}

\item{gattaca_dir}{Gattaca base directory. Default=''/data/apache/htdocs/genome_browser/lplab'}

\item{gattaca_folder_hub}{Folder in the server where to save hub directory}

\item{gattaca_user}{User of gattaca server ('user@gattaca')}
}
\value{
The hub.txt path of the rsync hub.
}
\description{
Rsync local hub to gattaca server and generates the html for uploading in ucsc page.
}
\examples{
path <- downloadUcschubgenerator()
hub_name <- 'example_hub_multiple'
path_local_hub <- file.path(path, 'hubs')
gattaca_folder_hub <- 'exampleHub'
gattaca_user <- 'msubirana@gattaca'
rsyncHub (hub_name = hub_name,
          gattaca_folder_hub = gattaca_folder_hub,
          gattaca_user = gattaca_user,
          path_local_hub = path_local_hub)
}
