% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackhubTrack.R
\name{trackhubTrack}
\alias{trackhubTrack}
\title{trackhubTrack}
\usage{
trackhubTrack(
  gattaca_html = "/data/apache/htdocs/genome_browser/lplab",
  gattaca_folder_hub,
  path_local_hub,
  hub_name = hub_name,
  assembly_database,
  track,
  quote_label = "",
  short_label = NULL,
  long_label = NULL,
  type = NULL,
  visibility = NULL,
  color = NULL,
  autoScale = NULL
)
}
\value{
Creates a folder with the basics (hub.txt and genomes.txt) for the generation of a UCSC hub (based on gattaca server).
}
\description{
Generates the trackDb file for every file in a defined path and copy the files into the local hub folder
in the defined assembly database.
}
\examples{
path_tracks <- path
hub_name <- 'example_hub_multiple'
path_local_hub <- file.path(path, 'hubs')
dir.create(path_local_hub)
assembly_database <- 'hg38'
gattaca_folder_hub <- 'exampleHub'
pattern_tracks <- '.vcf.gz$'
tracks <- list.files(path_tracks,
                     pattern = pattern_tracks,
                     full.names = T)
quote_label = 'vcf_examples'
type <- 'vcfTabix'
visibility <- 'dense'
color <- '0, 0, 0'

for(track in tracks){
  trackhubTrack(gattaca_folder_hub = gattaca_folder_hub,
                path_local_hub = path_local_hub,
                hub_name = hub_name,
                assembly_database = assembly_database,
                track = track,
                short_label = short_label,
                long_label = long_label,
                type = type,
               visibility = visibility,
                color = color,
                quote_label = quote_label)
}
}
