% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubGenerator.R
\name{hubGenerator}
\alias{hubGenerator}
\title{R parser for UCSC truckhub creation}
\usage{
hubGenerator(
  pathHub,
  hubName,
  hubShortLabel,
  hubLongLabel,
  emailAddress,
  descriptionUrl = NULL,
  assemblyDatabase,
  gattacaHtml,
  gattacaFolderHub
)
}
\arguments{
\item{pathHub}{Trackhub path}

\item{hubName}{Top-level name of the hub}

\item{hubShortLabel}{Short label for the hub, alias for UCSC parameter shortLabel}

\item{hubLongLabel}{Long label for the hub, alias for UCSC parameter longLabel}

\item{emailAddress}{Email that will be provided in the hub for contact info}

\item{descriptionUrl}{URL to HTML page with a description of the hub's contents}

\item{assemblyDatabase}{Reference genome used for alignment}

\item{gattacaHtml}{Gattaca html}

\item{gattacaFolderHub}{Folder in the server where to save hub directory}

\item{gattacaDir}{Gattaca base directory}

\item{gattacaUser}{User of gattaca server (user@gattaca)}
}
\description{
R parser for UCSC truckhub creation
}
